+++ packages/datatrak-web/src/api/mutations/useLogin.ts
@@ -24,7 +24,8 @@ export const useLogin = () => {
   const navigate = useNavigate();
   const location = useLocation();
 
-  const from = hasFrom(location.state) ? location.state.from : undefined;
+  // const from = hasFrom(location.state) ? location.state.from : undefined;
+  const from = location.state && location.state.from;
 
   return useMutation<any, Error, LoginCredentials, unknown>(
     ({ email, password }: LoginCredentials) => {
