+++ packages/datatrak-web/src/features/Survey/Components/index.ts
@@ -7,3 +7,4 @@ export { SurveyQuestion } from './SurveyQuestion';
 export { SurveyQuestionGroup } from './SurveyQuestionGroup';
 export { SurveyToolbar } from './SurveyToolbar';
 export * from './SurveySideMenu';
+export { SurveyReviewSection } from './SurveyReviewSection';
