+++ packages/datatrak-web/src/features/Survey/Screens/SurveyReviewScreen.tsx
@@ -5,10 +5,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Typography } from '@material-ui/core';
-import { QuestionType } from '@tupaia/types';
-import { formatSurveyScreenQuestions, getSurveyScreenNumber } from '../utils';
-import { useSurveyForm } from '../SurveyContext';
-import { SurveyQuestionGroup } from '../Components';
+import { SurveyReviewSection } from '../Components';
 import { ScrollableBody } from '../../../layout';
 
 const Header = styled.div`
@@ -20,39 +17,6 @@ const Header = styled.div`
   }
   ${({ theme }) => theme.breakpoints.up('md')} {
     padding: 1.375rem 2.75rem;
-  }
-`;
-
-const Section = styled.section`
-  padding: 1rem 0;
-  &:first-child {
-    padding-top: 0;
-  }
-`;
-
-const SectionHeader = styled(Typography).attrs({
-  variant: 'h3',
-})`
-  font-size: 1rem;
-  font-weight: ${({ theme }) => theme.typography.fontWeightMedium};
-  margin-bottom: 1rem;
-  ${({ theme }) => theme.breakpoints.up('sm')} {
-    font-size: 1.125rem;
-  }
-`;
-
-const Fieldset = styled.fieldset.attrs({
-  disabled: true,
-})`
-  border: none;
-  margin: 0;
-  padding: 0;
-  input,
-  label,
-  button,
-  .MuiInputBase-root,
-  link {
-    pointer-events: none;
   }
 `;
 
@@ -71,26 +35,6 @@ const PageDescription = styled(Typography)`
 `;
 
 export const SurveyReviewScreen = () => {
-  const { visibleScreens } = useSurveyForm();
-  if (!visibleScreens || !visibleScreens.length) return null;
-
-  // split the questions into sections by screen so it's easier to read the long form
-  const questionSections = visibleScreens.map(screen => {
-    const { surveyScreenComponents } = screen;
-    const screenNumber = getSurveyScreenNumber(visibleScreens, screen);
-    const heading = surveyScreenComponents[0].text;
-    const firstQuestionIsInstruction = surveyScreenComponents[0].type === QuestionType.Instruction;
-
-    // if the first question is an instruction, don't display it, because it will be displayed as the heading
-    const questionsToDisplay = firstQuestionIsInstruction
-      ? surveyScreenComponents.slice(1)
-      : surveyScreenComponents;
-    return {
-      heading,
-      questions: formatSurveyScreenQuestions(questionsToDisplay, screenNumber),
-    };
-  });
-
   return (
     <>
       <Header>
@@ -101,14 +45,7 @@ export const SurveyReviewScreen = () => {
         </PageDescription>
       </Header>
       <ScrollableBody>
-        <Fieldset>
-          {questionSections.map(({ heading, questions }, index) => (
-            <Section key={index}>
-              <SectionHeader>{heading}</SectionHeader>
-              <SurveyQuestionGroup questions={questions} />
-            </Section>
-          ))}
-        </Fieldset>
+        <SurveyReviewSection />
       </ScrollableBody>
     </>
   );
