+++ packages/datatrak-web/src/views/LandingPage/SurveyResponsesSection.tsx
@@ -38,24 +38,26 @@ export const SurveyResponsesSection = () => {
       <SectionHeading>My recent responses</SectionHeading>
       <ScrollBody>
         {isSuccess && recentSurveyResponses?.length > 0 ? (
-          recentSurveyResponses.map(({ id, surveyName, dataTime, entityName, countryName }) => (
-            <Tile
-              key={id}
-              title={surveyName}
-              text={entityName}
-              to={`/survey/1/survey-response/${id}`}
-              tooltip={
-                <>
-                  {surveyName}
-                  <br />
-                  {entityName}
-                </>
-              }
-              Icon={SurveyTickIcon}
-            >
-              {countryName}, {shortDate(dataTime)}
-            </Tile>
-          ))
+          recentSurveyResponses.map(
+            ({ id, surveyName, surveyCode, dataTime, entityName, countryName }) => (
+              <Tile
+                key={id}
+                title={surveyName}
+                text={entityName}
+                to={`/survey/${surveyCode}/response/${id}`}
+                tooltip={
+                  <>
+                    {surveyName}
+                    <br />
+                    {entityName}
+                  </>
+                }
+                Icon={SurveyTickIcon}
+              >
+                {countryName}, {shortDate(dataTime)}
+              </Tile>
+            ),
+          )
         ) : (
           <Typography variant="body2" color="textSecondary">
             No recent surveys responses to display
